Attempt a basic refactor. Give up.
